As suggested by Anne at https://issues.apache.org/jira/browse/OFBIZ-4675 reverts changes from https://issues.apache.org/jira/browse/OFBIZ-4381 r1167510
The fix applied in OFBIZ-4381 was to copy DemoShipping.xml from the ecommerce component (where it is loaded as DEMO data), and add it to commonext component as SEED data.
IMO this data should not be part of SEED data, as not all users will use this data. Also, the file should not be duplicated.
I suggest reverting trunk r1167510, R11.04 r1167513, R10.04 r1167514.
I suspect the correct fix for OFBIZ-4381 is either for the user to load the demo data, or to configure the carriers before trying to create a ProductStore.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/commonext@1240847 13f79535-47bb-0310-9956-ffa450edef68