Implements "Allow to edit ship groups contents after and order has been created" https://issues.apache.org/jira/browse/OFBIZ-5761
From an UI perspective this adds a "View ship group by order item" button which allows to create new ship groups by order items and to move the quantity of the order item between ship groups.
This is the adaptation of the "oisg-management" addon for R12.04 to the trunk. Thanks to Nicolas and Leila from Nereides for their help

For more information refer to the Jira issue.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/commonext@1643077 13f79535-47bb-0310-9956-ffa450edef68